Readded the updated unity packages of both unity chan project and Live2D Hiyori project.
Updates of Unity side

1. Separated the TCP related function from the controller script of both characters.

Although it makes previous tutorials depreciated, it allows user to open multiple instances of the Unity project, which each of them can have a different model.

Moreover, it gives control to the user when to start and stop the TCP server in Unity.

2. Completed UI system in Unity.

Two panels are added to the Unity application.

First is the character control panel. It controls how the model responses with the raw values sent from python, like the threshold of blinking and mouth movement. User can also save the values which will be loaded automatically when he/she opens the application again.

Second is the TCP server panel. Two buttons are provided to start and stop the Unity server at any time. Also, a text edit component is provided to allow user to input the port at their own wish. The range of the port value can be customized in the TCPServer.cs script. The custom port feature allows multiple instances of the application can be run and connected with their own instance of python main.py session.

The UI panel will automatically resize thanks to the AdjustGridLayoutCellSize.cs script.

3. Application aspect ratio adjustment.

The aspect ratio is adjusted to 9:16 portrait mode to show the full body of the models, instead of just showing the upper body of the model.

These changes can be experienced when importing the packages in UnityAssets/Packages/

The size of each packages are less than 100 MB, which should raise no problem with the git lfs and cloning the project to the local.